requirements.txt
arrow==1.3.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.8.2
crontab==1.0.4
cryptography==44.0.3
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-import-export==4.3.7
django-modeltranslation==0.19.14
django-redis==5.4.0
django-rq==3.0.1
django-templated-mail==1.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
ffmpeg-python==0.2.0
freezegun==1.5.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
moviepy==1.0.3
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rq==2.3.3
rq-scheduler==0.13.1
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tablib==3.8.0
times==0.7
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
